refactor: use constant reference file;
